Reorganize tests in projector_test
